+++ ui/component/claimListDiscover/view.jsx
@@ -31,6 +31,7 @@ type Props = {
   claimSearchByQuery: {
     [string]: Array<string>,
   },
+  claimSearchByQueryLastPageReached: { [string]: boolean },
   hiddenUris: Array<string>,
   hiddenNsfwMessage?: Node,
   channelIds?: Array<string>,
@@ -63,6 +64,7 @@ function ClaimListDiscover(props: Props) {
   const {
     doClaimSearch,
     claimSearchByQuery,
+    claimSearchByQueryLastPageReached,
     tags,
     defaultTags,
     loading,
@@ -297,6 +299,7 @@ function ClaimListDiscover(props: Props) {
   const hasMatureTags = tagsParam && tagsParam.split(',').some(t => MATURE_TAGS.includes(t));
   const claimSearchCacheQuery = createNormalizedClaimSearchKey(options);
   const claimSearchResult = claimSearchByQuery[claimSearchCacheQuery];
+  const claimSearchResultLastPageReached = claimSearchByQueryLastPageReached[claimSearchCacheQuery];
 
   const [prevOptions, setPrevOptions] = useState(null);
 
@@ -468,10 +471,7 @@ function ClaimListDiscover(props: Props) {
 
   function handleScrollBottom() {
     if (!loading && infiniteScroll) {
-      if (claimSearchResult && claimSearchResult.length % CS.PAGE_SIZE === 0) {
-        // Only increment the page if the current page is full. A partially-filled page probably
-        // indicates "no more search results" (at least based on my testing). Gating this prevents
-        // incrementing the page when scrolling upwards.
+      if (claimSearchResult && !claimSearchResultLastPageReached) {
         setPage(page + 1);
       }
     }